Allow custom values for plant community action type
